Upload de imagens a pelo JSON no formato base64. Adição de end-point para exclusão da imagem. Ajuste nos testes adicionando os Claims

+++ fase-01/back-end/Controllers/UserController.cs
@@ -4,6 +4,7 @@ using fase_01.application.interfaces;
 using fase_01.application.dtos;
 using fase_01.application.mappings;
 using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace fase_01.Controllers;
 
@@ -35,12 +36,21 @@ public class UserController : ControllerBase
     [HttpGet("{id:int}")]
     [Authorize]
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType<UserDto>(StatusCodes.Status200OK)]
     public async Task<IActionResult> GetById(int id)
     {
+        // obtém o ID do usuário conectado através das Claims
+        var currentUserIdClaims = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (!int.TryParse(currentUserIdClaims, out var currentUserId))
+            return Forbid();
+        if (!(currentUserId == id || User.IsInRole("Admin")))
+            return Forbid();
+
         var entity = await _userRepository.GetByIdAsync(id);
         if (entity == null) return NotFound();
+
         return Ok(entity.ToDto());
     }
 
@@ -52,20 +62,35 @@ public class UserController : ControllerBase
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public async Task<IActionResult> Update(int id, UserDto dto)
+    public async Task<IActionResult> Update(int id, [FromBody] UserDto dto)
     {
 
-        if (ModelState.IsValid)
+        if (!ModelState.IsValid)
             return BadRequest(ModelState);
 
         var existingEntity = await _userRepository.GetByIdAsync(dto.Id);
         if (existingEntity == null) return NotFound();
 
+        // Obtém o ID do usuário conectado via Claim
+        var currentUserIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        int.TryParse(currentUserIdClaim, out var currentUserId);
+
+        // Se o usuário logado estiver tentando remover o seu próprio status de Admin
+        if (currentUserId == id && existingEntity.Admin && !dto.Admin)
+        {
+            var allUsers = await _userRepository.ListAllAsync();
+            bool hasOtherAdmin = allUsers.Any(u => u.Admin && u.Id != id);
+
+            if (!hasOtherAdmin)
+            {
+                ModelState.AddModelError("Admin", "Não é possível remover a permissão de administrador do seu próprio usuário enquanto você for o único administrador do sistema.");
+                return BadRequest(ModelState);
+            }
+        }
+
         var entity = dto.ToEntity(existingEntity);
         await _userRepository.UpdateAsync(entity);
-
-        if (dto.Photo != null && dto.Photo.Length > 0)
-            await _photoService.SaveUserPhotoAsync(entity.Id, dto.Photo);
+        await _photoService.SaveUserPhotoAsync(entity.Id, dto.PhotoBase64);
 
         return NoContent();
 
@@ -81,6 +106,26 @@ public class UserController : ControllerBase
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     public async Task<IActionResult> Delete(int id)
     {
+        var existingEntity = await _userRepository.GetByIdAsync(id);
+        if (existingEntity == null) return NoContent();
+
+        // Obtém o ID do usuário conectado via Claim
+        var currentUserIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        int.TryParse(currentUserIdClaim, out var currentUserId);
+
+        // Se o usuário logado estiver tentando remover o seu próprio status de Admin
+        if (currentUserId == id && existingEntity.Admin)
+        {
+            var allUsers = await _userRepository.ListAllAsync();
+            bool hasOtherAdmin = allUsers.Any(u => u.Admin && u.Id != id);
+
+            if (!hasOtherAdmin)
+            {
+                ModelState.AddModelError("Admin", "Não é possível excluir o seu usuário enquanto você for o único administrador do sistema.");
+                return BadRequest(ModelState);
+            }
+        }
+
         await _userRepository.DeleteAsync(id);
         return NoContent();
     }
@@ -108,7 +153,33 @@ public class UserController : ControllerBase
             return NotFound();
 
         var bytes = photo.Thumbnail ?? photo.Image;
-        return File(bytes, photo.ContentType);
+        return File(bytes, "image/jpeg"); // thumbnail está sendo salvo como jpeg
+    }
+
+    [HttpDelete("photo/{id:int}")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    public async Task<IActionResult> DeletePhoto(int id)
+    {
+        var existingEntity = await _userRepository.GetByIdAsync(id);
+        if (existingEntity == null) return NoContent();
+
+        // Obtém o ID do usuário conectado via Claim
+        var currentUserIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        int.TryParse(currentUserIdClaim, out var currentUserId);
+
+        // Se for o próprio usuário ou um administrador tudo bem.
+        if (!(currentUserId == id || existingEntity.Admin))
+        {
+            ModelState.AddModelError("DefaultErrorMessage", "Apenas o próprio usuário ou um administrador por excluir a foto");
+            return BadRequest(ModelState);
+        }
+
+
+        await _photoService.DeleteUserPhotoAsync(id);
+        return NoContent();
     }
 
     #endregion
